Minor changes to updated permissions as well as a missed conflict resolution.

+++ x/ethbridge/types/claim.go
@@ -41,7 +41,7 @@ func NewEthBridgeClaim(
 		Amount:                amount,
 		ClaimType:             claimType,
 		TokenName:             tokenName,
-		Decimals:              int32(decimals),
+		Decimals:              int64(decimals),
 		DenomHash:             denomHash,
 	}
 }
